# openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_SessionResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SessionResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[SessionResponse]",
        "type": "object"
      },
...
      "Envelope_TokenResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TokenResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[TokenResponse]",
        "type": "object"
      },
...
      "Envelope_UserResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[UserResponse]",
        "type": "object"
      },
      "Envelope_UsernameCheckResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UsernameCheckResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[UsernameCheckResponse]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "도메인 에러 코드. 분기는 message 가 아니라 이 값으로 한다",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "OAuthGoogleRequest": {
        "properties": {
          "oauth_code": {
            "minLength": 1,
            "title": "Oauth Code",
            "type": "string"
          }
        },
        "required": [
          "oauth_code"
        ],
        "title": "OAuthGoogleRequest",
        "type": "object"
      },
      "OAuthGoogleResponse": {
        "description": "기존 사용자는 세션, 신규 사용자는 가입 티켓. `meta.is_new_user` 로 가른다.",
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionResponse"
              },
              {
                "$ref": "#/components/schemas/SignupTicketResponse"
              }
            ],
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/OAuthMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "OAuthGoogleResponse",
        "type": "object"
      },
      "OAuthMeta": {
        "properties": {
          "is_new_user": {
            "description": "true 면 data 는 가입 티켓이다 — `/auth/signup` 으로 가입을 마친다",
            "title": "Is New User",
            "type": "boolean"
          }
        },
        "required": [
          "is_new_user"
        ],
        "title": "OAuthMeta",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/TokenResponse"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "tokens"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SignupRequest": {
        "properties": {
          "role": {
            "enum": [
              "writer",
              "aspiring_writer",
              "reader"
            ],
            "title": "Role",
            "type": "string"
          },
          "signup_ticket": {
            "minLength": 1,
            "title": "Signup Ticket",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "signup_ticket",
          "role"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "SignupTicketResponse": {
        "description": "신규 사용자. 계정은 아직 만들지 않았다 (AUTH 계약 §4.2).",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "signup_ticket": {
            "title": "Signup Ticket",
            "type": "string"
          }
        },
        "required": [
          "signup_ticket",
          "email"
        ],
        "title": "SignupTicketResponse",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "auth_provider": {
            "title": "Auth Provider",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "enum": [
              "writer",
              "aspiring_writer",
              "reader"
            ],
            "title": "Role",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "email",
          "role",
          "auth_provider",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserUpdate": {
        "properties": {
          "role": {
            "enum": [
              "writer",
              "aspiring_writer",
              "reader"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "UserUpdate",
        "type": "object"
      },
      "UsernameCheckResponse": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "available"
        ],
        "title": "UsernameCheckResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "성공 응답은 `{data, meta}`, 실패는 `{error: {code, message, details}}` 다. 분기는 HTTP 상태가 아니라 `error.code` 로 한다 — 같은 상태 코드에 코드가 여럿이다.\n\n입력 검증 실패는 전부 **400 `INVALID_INPUT`** 이며 `details.fields` 에 위치가 온다. 인증은 `Authorization: Bearer <access_token>`. 목록은 커서 페이지네이션이다: `meta.next_cursor` 를 다음 요청의 `cursor` 로 넘기고, null 이면 끝이다.",
    "title": "Prolog (StoryForge) API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/logout": {
      "post": {
        "description": "리프레시 토큰을 폐기한다. 이미 폐기됐거나 모르는 토큰이어도 204 다.\n\n액세스 토큰은 만료될 때까지 유효하므로 클라이언트에서 지운다.",
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          }
        },
        "summary": "로그아웃한다",
        "tags": [
          "AUTH"
        ]
      }
    },
    "/v1/auth/oauth/google": {
      "post": {
        "description": "Google 이 준 authorization code 를 넘긴다. code 는 1회용이라 이 호출은 한 번만 성공한다.\n\n- 이미 가입한 사용자: `data` 는 사용자와 토큰, `meta.is_new_user = false`.\n- 처음 온 사용자: 계정을 만들지 않고 `data` 에 `signup_ticket`(10분 유효)과 이메일을 준다.\n  `meta.is_new_user = true`. 아이디·역할을 받아 `POST /auth/signup` 으로 가입을 마친다.",
        "operationId": "oauth_google_v1_auth_oauth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthGoogleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthGoogleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "502": {
            "content": {
              "application/json": {
                "examples": {
                  "OAUTH_PROVIDER_ERROR": {
                    "summary": "소셜 로그인 공급자 응답이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "OAUTH_PROVIDER_ERROR",
                        "details": {},
                        "message": "소셜 로그인 공급자 응답이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `OAUTH_PROVIDER_ERROR` — 소셜 로그인 공급자 응답이 올바르지 않습니다"
          }
        },
        "summary": "Google 로 로그인하거나 가입을 시작한다",
        "tags": [
          "AUTH"
        ]
      }
    },
    "/v1/auth/signup": {
      "post": {
        "description": "`POST /auth/oauth/google` 이 준 `signup_ticket` 에 아이디와 역할을 붙여 계정을 만든다.\n\n성공하면 바로 로그인된 상태(토큰 포함)를 돌려준다. 티켓은 한 번만 쓸 수 있다.",
        "operationId": "signup_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SessionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  },
                  "USERNAME_REQUIRED": {
                    "summary": "아이디가 필요합니다",
                    "value": {
                      "error": {
                        "code": "USERNAME_REQUIRED",
                        "details": {},
                        "message": "아이디가 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다\n- `USERNAME_REQUIRED` — 아이디가 필요합니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "SIGNUP_TICKET_INVALID": {
                    "summary": "가입 티켓이 만료되었거나 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "SIGNUP_TICKET_INVALID",
                        "details": {},
                        "message": "가입 티켓이 만료되었거나 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `SIGNUP_TICKET_INVALID` — 가입 티켓이 만료되었거나 올바르지 않습니다"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "USERNAME_TAKEN": {
                    "summary": "이미 사용 중인 아이디입니다",
                    "value": {
                      "error": {
                        "code": "USERNAME_TAKEN",
                        "details": {},
                        "message": "이미 사용 중인 아이디입니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `USERNAME_TAKEN` — 이미 사용 중인 아이디입니다"
          }
        },
        "summary": "가입을 마친다",
        "tags": [
          "AUTH"
        ]
      }
    },
    "/v1/auth/token/refresh": {
      "post": {
        "description": "리프레시 토큰을 새 액세스·리프레시 토큰 쌍으로 바꾼다.\n\n**로테이션이다.** 넘긴 리프레시 토큰은 즉시 폐기되므로 응답의 새 토큰으로 교체해 저장한다.\n폐기된 토큰을 다시 보내면 `REFRESH_TOKEN_INVALID` 다.",
        "operationId": "refresh_v1_auth_token_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TokenResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "REFRESH_TOKEN_INVALID": {
                    "summary": "리프레시 토큰이 유효하지 않습니다",
                    "value": {
                      "error": {
                        "code": "REFRESH_TOKEN_INVALID",
                        "details": {},
                        "message": "리프레시 토큰이 유효하지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `REFRESH_TOKEN_INVALID` — 리프레시 토큰이 유효하지 않습니다"
          }
        },
        "summary": "액세스 토큰을 재발급한다",
        "tags": [
          "AUTH"
        ]
      }
    },
...
    "/v1/users/check-username": {
      "get": {
        "description": "가입 화면용. 로그인하지 않아도 부를 수 있다.\n\n`available = true` 여도 가입 전에 다른 사람이 먼저 가져갈 수 있다 — 그때 가입은\n`USERNAME_TAKEN` 으로 실패한다.",
        "operationId": "check_username_v1_users_check_username_get",
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "required": true,
            "schema": {
              "maxLength": 30,
              "minLength": 1,
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UsernameCheckResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          }
        },
        "summary": "아이디 사용 가능 여부를 확인한다",
        "tags": [
          "AUTH"
        ]
      }
    },
    "/v1/users/me": {
      "get": {
        "description": "액세스 토큰의 사용자 정보.",
        "operationId": "me_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UserResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "USER_NOT_FOUND": {
                    "summary": "사용자를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "USER_NOT_FOUND",
                        "details": {},
                        "message": "사용자를 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `USER_NOT_FOUND` — 사용자를 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내 정보를 조회한다",
        "tags": [
          "AUTH"
        ]
      },
      "patch": {
        "description": "바꿀 수 있는 것은 `role`(writer · aspiring_writer · reader) 하나다. 아이디는 바꿀 수 없다.",
        "operationId": "update_me_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UserResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "USER_NOT_FOUND": {
                    "summary": "사용자를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "USER_NOT_FOUND",
                        "details": {},
                        "message": "사용자를 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `USER_NOT_FOUND` — 사용자를 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내 역할을 바꾼다",
        "tags": [
          "AUTH"
        ]
      }
    }
  }
}